[POC] Use okhttp for Android network layer
This allows us to enable HTTP/2 support

Relates-To: DATASDK-66
